refactor: move the profile field table into its own file
