+++ test/features/dive_log/domain/codecs/profile_series_field_table_test.dart
@@ -2,6 +2,7 @@ import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:submersion/core/database/database.dart';
+import 'package:submersion/features/dive_log/domain/codecs/profile_field_table.dart';
 import 'package:submersion/features/dive_log/domain/codecs/profile_series_codec.dart';
 
 /// Columns of `dive_profiles` that identify the series, not the sample.
